Removed an unnecessary duplicate class
and moved code into an existing file to resolved CS0436 warning
